+++ src/main/java/jcrapi2/api/intern/clans/members/ClanMembersResponse.java
@@ -16,6 +16,6 @@
  */
 package jcrapi2.api.intern.clans.members;
 
-import jcrapi2.common.PaginationResponse;
+import supercell.api.wrapper.essentials.common.PaginationResponse;
 
 public class ClanMembersResponse extends PaginationResponse<Member> {}