+++ src/main/java/jcrapi2/api/intern/globaltournaments/GlobalTournamentsResponse.java
@@ -16,6 +16,6 @@
  */
 package jcrapi2.api.intern.globaltournaments;
 
-import jcrapi2.common.PaginationResponse;
+import supercell.api.wrapper.essentials.common.PaginationResponse;
 
 public class GlobalTournamentsResponse extends PaginationResponse<GlobalTournament> {}