+++ src/main/java/jcrapi2/api/intern/locations/LocationApi.java
@@ -16,7 +16,6 @@
  */
 package jcrapi2.api.intern.locations;
 
-import jcrapi2.api.Api;
 import jcrapi2.api.intern.locations.info.LocationRequest;
 import jcrapi2.api.intern.locations.info.LocationResponse;
 import jcrapi2.api.intern.locations.rankings.clan.ClanRankingsRequest;
@@ -39,6 +38,8 @@ import jcrapi2.api.intern.locations.seasons.global.rankings.pathoflegend.player.
 import jcrapi2.api.intern.locations.seasons.global.rankings.pathoflegend.player.TopPlayerPathOfLegendRankingsResponse;
 import jcrapi2.api.intern.locations.seasons.global.rankings.tournament.TopPlayerTournamentRankingsRequest;
 import jcrapi2.api.intern.locations.seasons.global.rankings.tournament.TopPlayerTournamentRankingsResponse;
+
+import supercell.api.wrapper.essentials.api.Api;
 
 import java.util.concurrent.Future;