+++ src/main/java/jcrapi2/api/intern/locations/LocationApiImpl.java
@@ -16,8 +16,6 @@
  */
 package jcrapi2.api.intern.locations;
 
-import jcrapi2.api.ApiContext;
-import jcrapi2.api.BaseApi;
 import jcrapi2.api.intern.locations.info.LocationRequest;
 import jcrapi2.api.intern.locations.info.LocationResponse;
 import jcrapi2.api.intern.locations.rankings.clan.ClanRankingsRequest;
@@ -40,6 +38,9 @@ import jcrapi2.api.intern.locations.seasons.global.rankings.pathoflegend.player.
 import jcrapi2.api.intern.locations.seasons.global.rankings.pathoflegend.player.TopPlayerPathOfLegendRankingsResponse;
 import jcrapi2.api.intern.locations.seasons.global.rankings.tournament.TopPlayerTournamentRankingsRequest;
 import jcrapi2.api.intern.locations.seasons.global.rankings.tournament.TopPlayerTournamentRankingsResponse;
+
+import supercell.api.wrapper.essentials.api.ApiContext;
+import supercell.api.wrapper.essentials.api.BaseApi;
 
 import java.util.concurrent.Future;