+++ src/main/java/jcrapi2/api/intern/locations/rankings/clan/ClanRankingsResponse.java
@@ -16,6 +16,6 @@
  */
 package jcrapi2.api.intern.locations.rankings.clan;
 
-import jcrapi2.common.PaginationResponse;
+import supercell.api.wrapper.essentials.common.PaginationResponse;
 
 public class ClanRankingsResponse extends PaginationResponse<ClanRanking> {}