+++ src/main/java/jcrapi2/api/intern/locations/seasons/global/TopPlayerLeagueSeasonsV2Request.java
@@ -16,9 +16,9 @@
  */
 package jcrapi2.api.intern.locations.seasons.global;
 
-import jcrapi2.common.PaginationRequest;
-
 import lombok.Builder;
+
+import supercell.api.wrapper.essentials.common.PaginationRequest;
 
 public class TopPlayerLeagueSeasonsV2Request extends PaginationRequest {