+++ src/main/java/jcrapi2/api/intern/locations/seasons/global/info/TopPlayerLeagueSeasonRequest.java
@@ -16,9 +16,9 @@
  */
 package jcrapi2.api.intern.locations.seasons.global.info;
 
-import jcrapi2.common.Request;
-
 import lombok.Builder;
+
+import supercell.api.wrapper.essentials.common.Request;
 
 import java.util.Map;